+++ phi.nix
@@ -23,11 +23,6 @@ in {
 
   services.polybar = {
     config = {
-      "bar/base" = {
-        # TODO: just prepend `nixpkgs` to `modules-right`
-        modules-right = mkForce "nixpkgs wireless ethernet fs memory date";
-      };
-
       "bar/left" = {
         "inherit" = "bar/base";
         monitor = "DisplayPort-1";
@@ -40,39 +35,6 @@ in {
       "bar/right" = {
         "inherit" = "bar/base";
         monitor = "DisplayPort-2";
-      };
-
-      "module/nixpkgs" = let 
-        latest-nixpkgs = (pkgs.writeScript "latest-nixpkgs" ''
-          #!${pkgs.stdenv.shell}
-          # get latest commit hash for channel
-          LATEST=$(${pkgs.curl}/bin/curl -s https://channels.nix.gsc.io/nixos-unstable/history | ${pkgs.coreutils}/bin/tail -n 1 | ${pkgs.coreutils}/bin/cut -d ' ' -f 1)
-          # get commit hash of currently running system
-          RUNNING=$(${pkgs.coreutils}/bin/cat /run/current-system/nixos-version | ${pkgs.coreutils}/bin/cut -d '.' -f 4)
-          export GIT_DIR=~/nixpkgs/.git
-          UPDATE_FOUND=false
-          # check if running commit exists in ~/nixpkgs
-          if ${pkgs.git}/bin/git cat-file -e $RUNNING^{commit} 2>/dev/null; then
-            if ! ${pkgs.git}/bin/git merge-base --is-ancestor $LATEST $RUNNING 2>/dev/null; then
-              UPDATE_FOUND=true
-            fi
-          else
-            # if running commit doesn't exist in ~/nixpkgs, it has to have come from a channel
-            # this means if $LATEST != $RUNNING, $LATEST must be newer
-            if [[ $(echo $LATEST | ${pkgs.coreutils}/bin/cut -c -7) != $(echo $RUNNING | ${pkgs.coreutils}/bin/cut -c -7) ]]; then
-              UPDATE_FOUND=true
-            fi
-          fi
-          if [[ $UPDATE_FOUND == "true" ]]; then
-            echo  $(echo $LATEST | ${pkgs.coreutils}/bin/cut -c -7)
-          else
-            echo  $(echo $RUNNING | ${pkgs.coreutils}/bin/cut -c -7)
-          fi
-        '');
-      in {
-        type = "custom/script";
-        exec = "${latest-nixpkgs}";
-        interval = 300;
       };
     };
     script = mkForce ''